pyproject.toml
[project]
name = "strands-eval"
version = "0.1.0"
description = "SQL Agent Evaluation Framework using Strands, DeepEval, and MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deepeval>=1.3.0",
    "strands-agents",
    "pytest>=8.0.0",
    "pytest-json-report>=1.5.0",
    "sqlparse>=0.5.0",
    "sqlglot>=25.0.0",
    "pandas>=2.0.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "agentcore>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "sql_agent: SQL agent evaluation tests",
    "slow: Marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--json-report",
    "--json-report-file=evaluation_results.json",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
